Updates script with proper relative path for mj-include files. Also adds console logs for the files changed
